validation

+++ src/main/java/pl/brzezinski/bookt/repository/SchemaTableRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface SchemaTableRepository extends JpaRepository<SchemaTable, Long> {
 
     List<SchemaTable> findAllByRestaurant(Restaurant restaurant);
-    List<SchemaTable> findAllByPlaces(Long places);
-    List<SchemaTable> findAllByRestaurantAndPlaces(Restaurant restaurant, Long places);
+    List<SchemaTable> findAllByPlaces(int places);
+    List<SchemaTable> findAllByRestaurantAndPlaces(Restaurant restaurant, int places);
     List<SchemaTable> findAllByRestaurantAndPlacesIsBetween(Restaurant restaurant, Long down, Long top);
 }
